Maya Projects update
So the boat file was done on maya 2018 and was restarted multiple times to try and get something decent looking.   Boat 3.0 is a 10 minute version of a boat which is not working out so well.  I am having problems with a extrusions crossing over other things and causing errors (visuallly).  My robot is coming along but I deleted a face and need to figure out how to get it back.