Robot UV Mapping Complete
So it took so long because I started with the Gun and auto mapped it to 100 pieces and then tried to sew them back together, this was a mistake.  I started again with the simple pieces and worked up to the hard ones. Also I am much better at Nonmanifold Geometry now.